moving new types into guardrail components
