+++ comps/__init__.py
@@ -16,9 +16,6 @@ from comps.cores.proto.docarray import (
     SearchedDoc,
     RerankedDoc,
     TextDoc,
-    FactualityDoc,
-    PIIDoc,
-    ScoreDoc,
     RAGASParams,
     RAGASScores,
     GraphDoc,
