+++ comps/cores/proto/docarray.py
@@ -12,22 +12,7 @@ from pydantic import Field, conint, conlist
 
 class TextDoc(BaseDoc):
     text: str
-
-
-class FactualityDoc(BaseDoc):
-    reference: str
-    text: str
-
-
-class PIIDoc(BaseDoc):
-    prompt: str
-    replace: bool
-    replace_method: str
-
-
-class ScoreDoc(BaseDoc):
-    score: float
-
+    
 
 class Base64ByteStrDoc(BaseDoc):
     byte_str: str
