+++ comps/guardrails/toxicity_harm/predictionguard/typedocs.py
@@ -1,0 +1,8 @@
+# Copyright (C) 2024 Intel Corporation
+# SPDX-License-Identifier: Apache-2.0
+
+from docarray import BaseDoc
+
+
+class ScoreDoc(BaseDoc):
+    score: float
